remove x data type cast

+++ vllm/ir/ops/layernorm.py
@@ -17,8 +17,7 @@ def rms_norm(
     variance = x_var.pow(2).mean(dim=-1, keepdim=True)
     x = x * torch.rsqrt(variance + epsilon)
     if weight is not None:
-        weight = weight.to(torch.float32)
-        x = x.to(torch.float32) * weight
+        x = x * weight.to(torch.float32)
     return x.to(orig_dtype)
 
 
@@ -59,8 +58,7 @@ def fused_add_rms_norm(
     variance = x_var.pow(2).mean(dim=-1, keepdim=True)
     x = x * torch.rsqrt(variance + epsilon)
     if weight is not None:
-        weight = weight.to(torch.float32)
-        x = x.to(torch.float32) * weight
+        x = x * weight.to(torch.float32)
     return x.to(orig_dtype), x_residual
